feat: Basic json serialization functionality
Basic json serialization code is added to enable on fly configuration.
Currently, application is set during startup and number and duration of
rtp flows are fixed from that point.

Aim is to use json in on fly configuration. As an introduction,
string key and string, array and object value pairs are implemented.

Resolves: #1