bug: Coverity warnings are resolved
Three high and four low coverity warnings are resolved
The high level warnings are due to possible negative array indexing after
an addition method. The negative index are avoided through a check.
The medium level varnings are due to uninitialized scalers, which are
initialised to zero as a resolution, and incorrect pointer increment
which is corrected.

Resolves: #5